# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF633 - Service Catalog Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Catalog API is one of Catalog Management API\
    \ Family. Service Catalog API goal is to provide a catalog of services. \n\n###\
    \ Resource \n- serviceCatalog\n\n### Operations\nService Catalog API performs\
    \ the following operations on the resource :\n- Retrieve an entity or a collection\
    \ of entities depending on filter criteria\n- Partial update of an entity (including\
    \ updating rules)\n- Create an entity (including default values and creation rules)\n\
    - Delete an entity (for administration purposes)\n- Manage notification of events"
  version: "4.0.0"
  title: "Service Catalog Management"
host: "serverRoot"
basePath: "/tmf-api/serviceCatalogManagement/v4/"
...
schemes:
- "https"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
...
  /serviceCandidate:
    get:
      tags:
      - "serviceCandidate"
      summary: "List or find ServiceCandidate objects"
      description: "This operation list or find ServiceCandidate entities"
      operationId: "listServiceCandidate"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceCandidate"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
    post:
      tags:
      - "serviceCandidate"
      summary: "Creates a ServiceCandidate"
      description: "This operation creates a ServiceCandidate entity."
      operationId: "createServiceCandidate"
      parameters:
      - in: "body"
        name: "serviceCandidate"
        description: "The ServiceCandidate to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceCandidate_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceCandidate"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
  /serviceCandidate/{id}:
    get:
      tags:
      - "serviceCandidate"
      summary: "Retrieves a ServiceCandidate by ID"
      description: "This operation retrieves a ServiceCandidate entity. Attribute\
        \ selection is enabled for all first level attributes."
      operationId: "retrieveServiceCandidate"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceCandidate"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceCandidate"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
    delete:
      tags:
      - "serviceCandidate"
      summary: "Deletes a ServiceCandidate"
      description: "This operation deletes a ServiceCandidate entity."
      operationId: "deleteServiceCandidate"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceCandidate"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
    patch:
      tags:
      - "serviceCandidate"
      summary: "Updates partially a ServiceCandidate"
      description: "This operation updates partially a ServiceCandidate entity."
      operationId: "patchServiceCandidate"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceCandidate"
        required: true
        type: "string"
      - in: "body"
        name: "serviceCandidate"
        description: "The ServiceCandidate to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceCandidate_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceCandidate"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
...
definitions:
...
  ServiceCandidate:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of this REST resource"
      href:
        type: "string"
        description: "Hyperlink reference to this REST resource"
      description:
        type: "string"
        description: "Description of this REST resource"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of this REST resource"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of the service\
          \ candidate."
      name:
        type: "string"
        description: "Name given to this REST resource"
      version:
        type: "string"
        description: "the version of service candidate"
      category:
        type: "array"
        description: "List of categories for this candidate"
        items:
          $ref: "#/definitions/ServiceCategoryRef"
      serviceSpecification:
        description: "The service specification implied by this candidate"
        $ref: "#/definitions/ServiceSpecificationRef"
      validFor:
        description: "The period for which this REST resource is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "ServiceCandidate is an entity that makes a service specification\
      \ available to a catalog. A\nServiceCandidate and its associated service specification\
      \ may be published - made visible - in any number of service catalogs, or in\
      \ none. One service specification can be composed of other service specifications."
    example:
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      serviceSpecification:
        targetServiceSchema:
          '@baseType': "@baseType"
          '@type': "@type"
          '@schemaLocation': "@schemaLocation"
        '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      description: "description"
      version: "version"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
      href: "href"
      category:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
  ServiceCandidate_Create:
    type: "object"
    required:
    - "name"
    properties:
      description:
        type: "string"
        description: "Description of this REST resource"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of this REST resource"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of the service\
          \ candidate."
      name:
        type: "string"
        description: "Name given to this REST resource"
      version:
        type: "string"
        description: "the version of service candidate"
      category:
        type: "array"
        description: "List of categories for this candidate"
        items:
          $ref: "#/definitions/ServiceCategoryRef"
      serviceSpecification:
        description: "The service specification implied by this candidate"
        $ref: "#/definitions/ServiceSpecificationRef"
      validFor:
        description: "The period for which this REST resource is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "ServiceCandidate is an entity that makes a service specification\
      \ available to a catalog. A\nServiceCandidate and its associated service specification\
      \ may be published - made visible - in any number of service catalogs, or in\
      \ none. One service specification can be composed of other service specifications.\n\
      Skipped properties: id,href"
    example:
      lifecycleStatus: "lifecycleStatus"
      '@baseType': "@baseType"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      serviceSpecification:
        targetServiceSchema:
          '@baseType': "@baseType"
          '@type': "@type"
          '@schemaLocation': "@schemaLocation"
        '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      description: "description"
      category:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
  ServiceCandidate_Update:
    type: "object"
    properties:
      description:
        type: "string"
        description: "Description of this REST resource"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of the service\
          \ candidate."
      name:
        type: "string"
        description: "Name given to this REST resource"
      version:
        type: "string"
        description: "the version of service candidate"
      category:
        type: "array"
        description: "List of categories for this candidate"
        items:
          $ref: "#/definitions/ServiceCategoryRef"
      serviceSpecification:
        description: "The service specification implied by this candidate"
        $ref: "#/definitions/ServiceSpecificationRef"
      validFor:
        description: "The period for which this REST resource is valid"
        $ref: "#/definitions/TimePeriod"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "ServiceCandidate is an entity that makes a service specification\
      \ available to a catalog. A\nServiceCandidate and its associated service specification\
      \ may be published - made visible - in any number of service catalogs, or in\
      \ none. One service specification can be composed of other service specifications.\n\
      Skipped properties: id,href,lastUpdate"
    example:
      lifecycleStatus: "lifecycleStatus"
      '@baseType': "@baseType"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      name: "name"
      serviceSpecification:
        targetServiceSchema:
          '@baseType': "@baseType"
          '@type': "@type"
          '@schemaLocation': "@schemaLocation"
        '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      description: "description"
      category:
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      - '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
...
  ServiceCategoryRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of category"
      href:
        type: "string"
        description: "Hypertext Reference of the category"
      name:
        type: "string"
        description: "Name of the category."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "The (service) category resource is used to group service candidates\
      \ in logical containers. Categories can contain other categories."
    example:
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
...
  ServiceSpecificationRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the service specification"
      href:
        type: "string"
        description: "Reference of the serviceSpecification"
      name:
        type: "string"
        description: "Name of the requiredServiceSpecification"
      version:
        type: "string"
        description: "Service specification version"
      targetServiceSchema:
        description: "A target service schema reference (TargetServiceSchemaRef).\
          \ The reference object to the schema and type of target service which is\
          \ described by service specification."
        $ref: "#/definitions/TargetServiceSchema"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "Service specification reference: ServiceSpecification(s) required\
      \ to realize a ProductSpecification."
    example:
      targetServiceSchema:
        '@baseType': "@baseType"
        '@type': "@type"
        '@schemaLocation': "@schemaLocation"
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
...
  TimePeriod:
    type: "object"
    properties:
      endDateTime:
        type: "string"
        format: "date-time"
        description: "End of the time period, using IETC-RFC-3339 format"
      startDateTime:
        type: "string"
        format: "date-time"
        description: "Start of the time period, using IETC-RFC-3339 format. If you\
          \ define a start, you must also define an end"
    description: "A period of time, either as a deadline (endDateTime only) a startDateTime\
      \ only, or both"
    example:
      startDateTime: "2000-01-23T04:56:07.000+00:00"
      endDateTime: "2000-01-23T04:56:07.000+00:00"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"